demo(android): unify dynamic-sky + environment + reflection-probes + post-processing into lighting-lab (#2239 Batch 2)
Consolidates the retired `dynamic-sky`, `environment`, `reflection-probes`,
and `post-processing` demos into a single `lighting-lab` entry with a 4-way
segmented-button toggle (mirrors #2239 Batch 1's 2D-in-3D merge):

- **Sky** (default) — DynamicSkyNode time-of-day sun + turbidity sliders.
- **Environment** — HDR IBL chips + IBL-intensity overrides.
- **Reflections** — ReflectionProbeNode local IBL zone (own scene/cameraNode).
- **Post-FX** — SSAO/MSAA/FXAA/dithering switches via rememberView + SideEffect.

Each sub-mode owns its own rememberEngine + loaders + SceneView so switching
tabs tears down the inactive section completely (prevents resource leaks).
All QA-day / #issue rationale comments preserved from the source demos.

Retired deep-link ids (`dynamic-sky`, `environment`, `reflection-probes`,
`post-processing`) keep working via DeepLinkRouter.DEMO_ID_ALIASES.

Tests: DemoInteractionTest.lightingLab_allTabs taps all 4 tabs;
DemoRenderingScreenshotTest.lightingLabDemo_default_state (golden
lightinglab_default). Stale 4 goldens removed.
