+++ changelog.d/2239-batch-2-lighting-lab.md
@@ -1,0 +1,2 @@
+<!-- category: Changed -->
+- **Samples catalog: unified Lighting Lab demo ([#2239](https://github.com/sceneview/sceneview/issues/2239)).** The retired `dynamic-sky`, `environment`, `reflection-probes`, and `post-processing` demos consolidated into a single `lighting-lab` entry with a segmented-button toggle between Sky (`DynamicSkyNode` time-of-day sun), Environment (HDR IBL switching), Reflections (`ReflectionProbeNode` local IBL zone), and Post-FX (SSAO / MSAA / FXAA / dithering) modes. Existing `sceneview://demo/dynamic-sky|environment|reflection-probes|post-processing` deep links keep working via `DEMO_ID_ALIASES`.
