+++ samples/android-demo/src/main/java/io/github/sceneview/demo/DeepLinkRouter.kt
@@ -104,6 +104,15 @@ internal object DeepLinkRouter {
         "image" to "two-d-in-three-d",
         "video" to "two-d-in-three-d",
         "billboard" to "two-d-in-three-d",
+        // #2239 Batch 2 — Lighting Lab consolidation. The retired `dynamic-sky`,
+        // `environment`, `reflection-probes`, and `post-processing` demos merged
+        // into `lighting-lab` with a segmented-button toggle. The unified demo
+        // opens on its default Sky tab (tab pre-selection from alias is a planned
+        // follow-up).
+        "dynamic-sky" to "lighting-lab",
+        "environment" to "lighting-lab",
+        "reflection-probes" to "lighting-lab",
+        "post-processing" to "lighting-lab",
     )
 
     fun parse(data: Uri?, registry: List<DemoEntry> = ALL_DEMOS): String? {
