--- samples/android-demo/src/main/res/values/strings_demo_dynamic_sky.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `dynamic-sky` demo (registered by
-     `DynamicSkyFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_dynamic_sky` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_dynamic_sky">Dynamic Sky</string>
-    <string name="demo_dynamic_sky_subtitle">Time-of-day sun simulation</string>
-    <string name="demo_dynamic_sky_desc">Time-of-day sun position with atmospheric scattering</string>
-</resources>
