--- samples/android-demo/src/main/res/values/strings_demo_environment.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `environment` demo (registered by
-     `EnvironmentFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_environment_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_environment_title">Environment Gallery</string>
-    <string name="demo_environment_subtitle">HDR environment switching</string>
-    <string name="demo_environment_gallery">Environment Gallery</string>
-    <string name="demo_environment_gallery_desc">Compare HDR environments on the same model side by side</string>
-</resources>
