--- samples/android-demo/src/main/res/values/strings_demo_post_processing.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Strings for the `post-processing` demo (registered by
-     `PostProcessingFragment.kt`). Android's resource merger
-     fans every `res/values/*.xml` file in, so `R.string.demo_post_processing_title` etc.
-     remains globally resolvable. Adding a new demo never needs to touch
-     this file or `strings.xml` — see #1870. -->
-<resources>
-    <string name="demo_post_processing_title">Post Processing</string>
-    <string name="demo_post_processing_subtitle">SSAO, anti-aliasing, and tone mapping</string>
-    <string name="demo_post_processing">Post-Processing</string>
-    <string name="demo_post_processing_desc">Bloom, SSAO, FXAA, tone mapping, vignette</string>
-</resources>
